Deal with Windows test failures
First if failed to compile because a constant (capi::L_CLONE) was of
type i32 rather than u32 as expected. Fortunately, it then gets cast at
i32, so i32 works out better for us.

I've changed the function `get_box` to expect an i32 rather than a u32,
and coerce `L_CLONE` to be i32.

I would have made `get_box` generic, but I couldn't see how to do
generic coercions.

A test that set the environment variable TESSDATA_PREFIX wasn't working.
As this isn't a core piece of functionality, I've set it not to run on
Windows.

Finally a doc test was relying on tesseract having a default location to
get trained data. This wasn't working as vcpkg tesseract doesn't have a
default location. As the test wasn't about not having default locations,
I set a default location.

https://github.com/houqp/leptess/issues/3

+++ src/leptonica.rs
@@ -103,9 +103,10 @@ impl Boxa {
         unsafe { (*self.raw).n as usize }
     }
 
-    pub fn get_box(&self, i: usize, flag: u32) -> Option<Box> {
+    pub fn get_box(&self, i: usize, flag: i32) -> Option<Box>
+    {
         unsafe {
-            let b = capi::boxaGetBox(self.raw, i as i32, flag as i32);
+            let b = capi::boxaGetBox(self.raw, i as i32, flag);
             if b.is_null() {
                 return None;
             }
@@ -148,7 +149,7 @@ impl Iterator for BoxaIterator {
             return None;
         }
 
-        let re = self.boxa.get_box(self.index, capi::L_CLONE);
+        let re = self.boxa.get_box(self.index, capi::L_CLONE as i32);
         self.index += 1;
 
         re
@@ -183,7 +184,7 @@ impl<'a> Iterator for BoxaRefIterator<'a> {
             return None;
         }
 
-        let re = self.boxa.get_box(self.index, capi::L_CLONE);
+        let re = self.boxa.get_box(self.index, capi::L_CLONE as i32);
         self.index += 1;
 
         re
